+++ SteamOrganizer/MVVM/View/Windows/ShowAuthenticatorWindow.xaml.cs
@@ -13,8 +13,6 @@ namespace SteamOrganizer.MVVM.View.Windows
             DataContext = new ShowAuthenticatorViewModel(account);
         }
 
-        private void deleteAuth_Click(object sender, RoutedEventArgs e)          => confirmDeleteNo.Visibility = confirmDeleteYes.Visibility = Visibility.Visible;
-        private void confirmDeleteHidden_Click(object sender, RoutedEventArgs e) => confirmDeleteYes.Visibility = confirmDeleteNo.Visibility = Visibility.Collapsed;
         private void codeCopyButton_Click(object sender, RoutedEventArgs e)      => Utils.Win32.Clipboard.SetText(guardBox.Text);
         private void BorderDragMove(object sender, MouseButtonEventArgs e)       => this.DragMove();
     }